Jukebox + Actions!
"So I noticed in the code a comment left by Sin_Ender for the jukebox permission that said "PSHHHHH this will probably never happen lmfao" so I had to lock in..."

That is what Im going to say in the next video 👍

Anyways, 2 new actions for starting (unpausing) and stoping the current music.

The entirety of the jukebox is done, however it doesnt save.